Add new option to exclude routes from being tracked
Sometimes it might be beneficial to not track visits to certain routes. 
With these changes, a different version of the tracker script will be 
used that allows excluding routes based on a data-attribute. Unless the 
excludeRoutes option is set in the config, the plugin works as before. 
It’s also possible to use both excludeRoutes and outboundLinkTracking at 
the same time.